// openapi/models/participant-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ParticipantResponse
 */
export interface ParticipantResponse {
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    'participantKey': string;
    /**
     * 
     * @type {boolean}
     * @memberof ParticipantResponse
     */
    'isEmailSent': boolean;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    'partytype': string;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    'role': string;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResponse
     */
    'id': number;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResponse
     */
    'eventid': number;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResponse
     */
    'partyid': number;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResponse
     */
    'userid': number;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    'firstname': string;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    'lastname': string;
    /**
     * 
     * @type {string}
     * @memberof ParticipantResponse
     */
    'email': string;
    /**
     * 
     * @type {number}
     * @memberof ParticipantResponse
     */
    'tenantid': number;
}